Board Draws On Page Load
In this commit, I altered the checkers.js to use a slightly
different form than it was previously, and I modified index
.html to start the javascript on <body> load. Also, I chan-
-ged a few other things stylistically, you guys can comment
on whether or not you approve of that and we can hopefully
establish a style guide. Mostly, I added whitespace in pla-
-ces I think there should be, including inside function
calls, in arithmetic expressions, and in initializing
structs. Let me know what you guys think!