Using WebGL instead of 2d context
I think the 2d context works really well, and has a lot of useful
functions already written for drawing to the page, however I
think it is important to compare and contrast technologies. If
we are at all interested in adding 3d features later on, a WebGL
context will be mandatory. So, I this is a branch where I will
complete FR-01 (and maybe more?) using WebGL, and we can compare
and contrast to the 2d context implementation of FR-01.