Draws a square
DESIGN DECISION:
I decided to move all variables which exist for the entirety of the start()
function to global scope. This is generally how game development is done,
and in the past I have found it easier to reason about (contrary to what you
might expect).
Also, I've added the initBuffers function for initializing
vertex buffer objects, and I added the draw function, for drawing
to the canvas.